build: Rename .conf to .build files
Search in whole directory for .build files and set IN and OUT variables
based on where the .build file was found. These variables can be overriden
in the .build file if wished so. For example if the build files are
kept in the root directory then IN and OUT variables must be set accordingly.